requirements.txt
Pillow
flask
flasgger
flask-cors
rembg
onnxruntime